Converted editor MP calculations to use doubles
this uses doubles rather than floats to ensure the tiniest of edits can also represent an increase in MP usage, so that there never needs to be a conversion from floats to doubles in this feature

+++ src/general/base_stage/EditorComponentBase.cs
@@ -171,7 +171,7 @@ public partial class EditorComponentBase<TEditor> : ControlWithInput, IEditorCom
     {
     }
 
-    public virtual void OnMutationPointsChanged(int mutationPoints)
+    public virtual void OnMutationPointsChanged(double mutationPoints)
     {
         throw new GodotAbstractMethodNotOverriddenException();
     }
